fix(firmware): release 0.8.1 with lazy Hall covers and NAND storage support

+++ firmware/espressif/main/host/controller/control_types.hpp
@@ -30,8 +30,8 @@ struct AppDescriptor final {
 struct CatalogSnapshot final {
     std::array<AppDescriptor, kMaxApps> apps{};
     uint32_t count{};
-    uint32_t store_total_bytes{};
-    uint32_t store_used_bytes{};
+    uint64_t store_total_bytes{};
+    uint64_t store_used_bytes{};
 };
 
 struct AppDiagnostic final {
